Resolves mod detail page issue
Fixes an issue where a mod which has failed to resolve to an SPT version can be viewed by anyone using the detail page link. Checks are now in place to ensure that only a mod author, moderator, or admin may view these mod detail pages, and if they do null checks are performed on the latestSptVersion properties.